🎨 Move parallel migrations to separate module
