+++ cmd/migrate/migrate.go
@@ -4,6 +4,10 @@ import (
 	"github.com/spf13/cobra"
 )
 
+const (
+	CmdBatchSize = "batch-size"
+)
+
 var MigrateCommand = &cobra.Command{
 	Use:   "migrate",
 	Short: "Run application level migration in parallel with the poller",
@@ -13,5 +17,6 @@ func init() {
 	MigrateCommand.AddCommand(
 		MigrationSetupIscnVersionTableCommand,
 		MigrationAddressPrefixCommand,
+		MigrationNftEventPriceCommand,
 	)
 }
